fix(web): Wait for bifrost to initialize before fetching the action view list from sqlite
This is the one case where we're rendering data before bifrost is initialized and before cold start finishes, which was causing a whole host of strange issues. Adding this flag to the context, allows the Apply button to render with a default 0 action count, and the action count updates when the query can finally run.