feat: added a cxx toolchain
We get our clang from https://github.com/systeminit/vfox-clang, a fork
of the main mise clang vfox effort that adds a symlink to clang++. We
use vfox over asdf so we are just grabbing binaries instead of compiling
clang, which can take hours.

There's a lot of dancing around here to ensure that all the things know
where clang and rustc are at the right time. We rely on wrapping calls
to the bins in mise and wrapper scripts that set PATH and other env
vars.

There are also a pile of changes to cargo.tomls for tracing. I don't
know why this is suddenly a problem here, but it failed with both buck2
and cargo build without it. Probably fixable once this gets a cleanup
pass.