fix(web): Remove incorrect defalt of all passing for qualifications
When we created a component in the new UI, we checked for failing or warning qualifications and if not found then we defaulted to `all passed`.

For new components, there will be no passing, no failing and no warning. So if we find that seet of groups then we default to `unknown` as the status is actually calculating the results.

This makes the flow that when you create a component that te qualifications show unknown until they work as expected

https://jam.dev/c/d382da8c-b166-4508-bbbc-449c69966b0c