Make streamed patches from edda a config setting instead of always on
Rather than always sending the patches for the front end both as
individual patches when each is built, and as a single batch after all
have been built, there is now a `SI_EDDA__STREAMING_PATCHES` setting to
toggle between the two behaviors. The default behavior is to send the
patches as a single batch, making the newer streaming behavior opt-in
for development & testing.