Remove running qualifications from new UI
This change removes running qualifications from the new UI. This is a
very old concept that predates qualifications on the prop tree (late
2022!) and has been moved around over the years. DVU is the canonical
source of prop tree functions in flight and qualification statuses are
reflections of the result of DVU.

As a consequence of this work, the domain naming has been standardized
across the newhotness Vue components.

An additional change here is that the two sets for action state for
sort by functionality have been combined under one interface to only
loop over the action list once.

Finally, there's now a comment clarifying why the "canUpgrade" variable
was created in a previous PR for the grid.

Signed-off-by: Nick Gerace <nick@systeminit.com>