Add ability to pin a component
This change adds the ability to pin a component. Pinning a component
results in two groups being shown underneath it: its incoming and
outgoing connecting components. The pinned component is placed at the
top of the grid and can be focused.

There are two ways to interact with a pinned component via the keyboard.
You can press "ENTER" to navigate to it or you can press "P" to unpin
it. Conversely, pinning a component can be done by pressing "P" or
using the component context menu. Note: the component context menu is
disabled for pinned components and "group by" is disabled when
pinning.

As a consequence of the primary change of this commit, component
qualification status badging and iconography has been moved into its
own Vue component. Not only does it make it reusable for the grid tile
and the pinned component card, but it also allows these details to be
available in the delete modal.

Other changes include the following:

- Add unfocusing when using group by, sort by and pinning due to invalid
  indexing
- Remove "Standby..." from the qualification helper text
- Remove leftover "Running qualifications" switch check from grid row
- Fix qualification status missing "unknown" for qualification view

Known missing feature: we did not add the pin icon on the upper right of
the grid tiles themselves.

Co-authored-by: Victor Bustamante <victor@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>