Edda: Try to find "old" snapshot if change set has moved
When doing the "stream compression" to collapse the pending requests
into a single batch of work, we check to see if the final
`to_snapshot_address` of the requests drained from the queue is the same
as the current snapshot address used by the change set.

Previously, if the final `to_snapshot_address` wasn't the same as the
snapshot address used by the change set we would fall back to a full
rebuild. While the "stream compression" significantly cut down on the
number of times we fell back to a full rebuild, this is still a race
condition that we're hitting fairly regularly in an active change set.

Now, if we see that the change set has moved on to a new snapshot
address, we'll try to retrieve the snapshot referenced in the final
`to_snapshot_address` and do an incremental build using that. If we
can't retrieve the snapshot from `to_snapshot_address` then we'll still
fall back to a full rebuild.

This moves the race condition from being between when we pull the final
request from the queue and when the rebaser does the next update to the
change set pointer, to being between when we pull the final request from
the queue and when the snapshot referenced in the `to_snapshot_address`
is removed through garbage collection.