Edda: Incrementally build MVs, even if change set has moved on to a new snapshot
We were falling back to a full rebuild if the current snapshot used by a
change set did not match the `to_snapshot_address` of the final request
that Edda pulled from the queue to ensure that we didn't end up with an
MV that ended up not being updated due to the change set pointer's
snapshot address and the `to_snapshot_address` getting out of sync.

This should only be a problem if the `to_snapshot_address` somehow got
_ahead_ of the snapshot referred to by the change set pointer, though.
If the `to_snapshot_address` gets ahead of the change set pointer, then
we'll be working with a list of Changes that the snapshot does not yet
have the associated information for, and when the change set pointer
catches up we will have already processed the Changes necessary to
trigger rebuilding the appropriate MVs, causing those MVs to remain out
of date. It should not be possible for the `to_snapshot_address` to get
ahead of the change set pointer, because the request containing the
`to_snapshot_address` is not sent by the rebaser until _after_ it has
updated the change set pointer.

The much more realistic (and common) case is for the change set pointer
to move forward between the time edda has pulled all pending requests
from the queue to get the final `to_snapshot_address` of the requests,
and when we check to see which snapshot the change set pointer is
referencing as the rebaser could have committed another set of updates
in that window.

This is not a problem, however, as all of the information related to the
Changes in the requests Edda has pulled from the queue will already be
in the snapshot referenced by the change set pointer. There would be
updated information in the workspace snapshot that we have not yet
been notified of through the Changes in the queued requests, but we
should see those Changes once we pull the next set of requests from the
queue.

For the change set pointer being ahead of the `to_snapshot_address` to
never be a problem, we will need to ensure that we have a contiguous
chain of the `from` & `to` snapshot addresses in the edda requests, even
across batches that are pulled from the queue. Within a batch, the `to`
must be the same as the following request's `from`, and across batches,
the final `to` of a batch must be the same as the initial `from` of the
following batch. If these chains are broken at any point, then it means
that we have missed a set of Changes and no longer have the information
necessary to know what the minimal set of MVs is that needs to be
rebuilt (and should fall back to a full rebuild).

We are already checking to make sure that the chain of `from` & `to`
snapshot addresses is unbroken within a batch of edda requests, but we
are not yet making sure the chain is unbroken across batches.