Fix component list and connections in Explore
This is "one of those" changes. Here's the deal: I originally set out to
just fix connections for pinning. As I clawed at every little bug fix,
the rabbit hole got deeper. I fixed the connections when pinning! Wait,
now the indexes are messed up. I fixed the indexes sometimes. Wait, it
still happens. Okay those look better, but... oh DEAR GOD THE COMPONENT
CONTEXT MENU IS FLYING OFF THE SCREEN.

There turned out to be something evil lurking within...the component
list, filtered component list, and grouped component list all did not
account for pinning. Sure, they accounted for views, but not pinning.
For indices, tabbing, component context menu and general ease of use,
you need your component list to be absolute. You cannot just drop
components from that list when grouping.

I combed all of Explore and now my brain is fried. It is happy, but
fried. Each section is now titled in the file and there is a clear split
between determining the absolute component list and performing
filtering, sorting and grouping of that list. More stuff in the logic
composables directory in the future? Likely yes.

Now, the original guts of this change are still here, despite all the
drama. That change fixes connections for pinning and makes connections
reusable for multiple contexts. Now, the connections panel and pinned
connections use the same underlying logic. Alongside that work, we now
bust the output count alongside the outgoing connections to ensure both
are working in tandem with one another.

All of these changes result in grid that is much more trustworthy and
reliable when working with pinning and connections.

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>