chore(edda): speed up external connection count in ComponentMv by 1000%
That much huh? Pretty much, yeah. Happy Friday ;)

This change implements a method `external_source_count()` on
`WorkspaceSnapshotSelector` which computes this count using lower level
abstractions on the snapshot and graph. As a result, the answer is
computed much more quickly and (it appears) without any additional
allocations.

There is a fair amount of refactoring to accompany this change, but sets
up the code for future changes in the same spirit.

Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>