Add system status in the new UI
This change adds system status to the new UI. Specifically, it uses
a "bucket" approach to handle different kinds of scenarios involving
status. It has auto-fade functionality and the ability to eject items
from the bucket that hit the timeout (which is hard-coded right now, but
could become dynamic to the kind of content in the bucket).

The bucket currently has two kinds of items in it right now: components
in flight (DVU roots) and "rainbow" items (MVs in flight). More kinds of
items are planned to be added in the future, like management functions.
As a result of these changes, the rainbow spinner has been removed.

The underlying mechanism opens the door to varying levels of granularity
of status and detail. Right now, the bucket is generic to all kinds of
status and the level of detail is low. However, the core logic should
be supple enough for us to go nuts.

This change also disables all modals related to the old UI when using
change set apply except for the one that indicates to users that
they've been moved to HEAD. Applying a change set should feel much
faster and the apply flow no longer relies on the old change sets
store. It now uses the new "useApi" composable.

One more thing too... the message saying "Because you are currently on
Head"... I changed "Head" to "HEAD" for consistency. It drove me nuts!

Finally, there is a known issue with this change: the new status icon
isn't perfectly centered to the change set buttons to the left of it.

Signed-off-by: Nick Gerace <nick@systeminit.com>