chore(dal): demote dependency graph calculating orphaned spans
This change demotes the following spans to `debug`:

- `list_topologically`
- `action.dependency_graph.for_workspace`

The `list_topologically` is currently orphaned an always has a child of
`action.dependency_graph.for_workspace` which accounts for 100% of the
parent span.

This trace combination was captured over 1.67M times in production
Honeycomb over the last 7 days. If future profiling is required in this
area we can dial the level back up.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>