chore(edda): demote "quiesced notified" log event to debug
This change demotes a log event when the `ChangeSetProcessorTask` begins
shutting down after detecting a quiescent period. This logic appears
much more trustworthy than when it was initially implemented and can be
safely demoted to a `debug` level.

This event was captured over 201K times in production Honeycomb over the
last 7 days.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>