First pass at minimizing Vec (re)allocations
As we learned in 7084431ffa, extra allocations are a death by a thousand
cuts for performance.

This focuses on the MV creation methods, specifically on uses
`Vec::new()`, and `vec![]` to pre-allocate (at least) the required
space, whenever possible using `Vec::with_capacity(final_entry_count)`.

In the case where we don't have a reasonable estimate of the final size
of the `Vec`, and we do know how many entries we're about to add,
reserve space for the batch before inserting using
`vec.reserve(new_item_count)`. This reduces the number of allocations
necessary to have space for all of the new entries down to a maximum of
one, instead of possibly having multiple re-allocations.