fix(dal): properly construct dvu graph without disjoint trees
Values discovered when walking object children, or when walking to the
parent, were not always connected to the components entire DVU
dependency tree when constructing the DVU graph. This typically led to
the resource and resource_value tree being treated as an independent set
of values, meaning it might run before, or after, the rest of the
component was constructed, meaning the view of the component might not
include the resource when *other* values that that depend on that
component's view had their function executed. The DVU graph construction
has been refactored a bit to ensure that there will be no "disjoint"
trees within a single component, or for inter-component dependencies.