Handle change set apply loading states
This change handles change set apply loading states and centralizes
logic to apply change sets. It also centralizes the logic for when the
ability to apply change sets should be disabled.

As a result of these changes, the approval flow modal apply button
follows the same loading and disabled states as the regular button and
vice versa. That fixed a bug where you can apply if you have the modal
open while someone applies the change set.

Signed-off-by: Nick Gerace <nick@systeminit.com>