fix(web): Remove HEAD warning on AddComponentModal
We don’t have any real way for a user to understand what HEAD actually means. 
Yes we connect to the docs but this is the only place we show this so it’s a confusing interaction.
We already force a changeset so we should continue with that thread only for now. This warning may 
even prevent a user from creating a component