Remove async_recursion from dal
While convenient, `async_recursion` introduces a fair amount of
overhead, and can end up blowing through the stack size limit much
faster than expected. Especially when recursing over user-defined data,
as is done in `AttributeValue::view`.

In addition to making `AttributeValue::view` iterative, instead of
recursive, this moves a lot of "graph level" code out of  the domain
objects and down into the graph level traits allowing for more direct
usage of the underlying graph, more re-use of code between the graph
level trait methods, and minimizing the construction of full
domain-objects only to pull specific pieces of information from them and
to gain access to the helper methods that require the full domain-object
instead of just an ID.