fix(web): support multiple workspaces
The bifrost websocket is workspace specific, so we need to connect
a new socket for every workspace that the user loads.

The sdf client injects the token when it is created, so we also need an
sdf client for every workspace that the user loads.

This pr ensures that we initialize a new socket and sdf client for every
workspace that the user loads, and that we re-initialize all the
necessary sockets and sdf clients when a new tab grabs the lock for the
database.