feat(sdf,web): use gzip compression when sending attribute update calls
This change adds a middleware to SDF's PUT `component/<id>/attributes`
route which requires the client to send the JSON body request compressed
using Gzip compression.

On the `app/web` side, the SDF API client now supports calls that must
be compressed, of which the "attribute update" call is the first and
only one.

The purpose of compressing this request body is in an effort to avoid a
Web Application Firewall (i.e. a WAF) from erroneously detecting
suspicious or malicious client content being sent to the server backend.
In this case some of the attribute value's contents are themselves JSON
objects and other elaborate data structures that have been detected and
rejected by AWS' WAF product.

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>