feat(web): add hashes to the frontend build
Adds two hashes to the frontend build:

`__COMMIT_HASH__`, the hash of HEAD when the frontend is built, and
`__SHARED_WORKER_HASH__`, the git sha of the shared_webworker.js file when
the frontend is built.

We can use the commit hash to confirm that a user is on the latest
version of the frontend (they can go to `/hashes`) to see the hashes,
and we will use the `__SHARED_WORKER_HASH__` to detect when the shared
webworker that is resident in the browser is out of sync with the
frontend code running in a tab.