fix(dal): When resetting an attribute value to a default prototype, first clear the existing value
Whenever an attribute value is set by a subscription, or has any other component specific prototype, upon resetting to the schema variant defined (default) prototype, if the default prototype is a dynamic function, we enqueue it in DVU for it to get updated.  However this means that value is temporarily stale until DVU completes, leading to an unexpected user experience.

So we go from: 
1. Value Set By Subscription = "vpc-12345"
2. Not subscription but value is still "vpc-12345"
3. DVU Runs and finishes, value is removed (or updated depending on the variant)

To: 
1. Value Set by Subscription = "vpc-12345"
2. No subscription, value = Null
3. DVU Runs and finishes, no change if it stays Null, or is updated.