feat(web): gracefully handle the deployment of a new shared webworker
If a new shared worker is detected, hotswap in our comlink to that
worker and update our remotes and bearer tokens.

Note that if the tab-local worker changes in breaking ways, or if the
interface contract between the tab and the shared worker changes, a
reload will still be required for other tabs to get the latest code. But
that should be handled by our CachedAppNotification component already.

To test this, load the app in a few tabs (for a real test, load more
than one workspace locally, with at least 3 tabs). Then go into
`heimdall.ts` and modify the `sharedWebWorkerName` var. Then reload in
just *one* tab. Ensure that all tabs work correctly and that updates
flow across the websocket(s).

Then open `chrome://inspect/#workers` or
`about:debugging#/runtime/this-firefox`. There should only be one shared
webworker. If you see two, reload the inspect page (it is not always
immediately reactive).

Note that after this is deployed, every tab will have to be closed to
support the new shared worker code, but only one final time.
Afterwards, we should gracefully detect new shared workers without
requiring all tabs to close.