feat(web): compute possible connections per component
We calculated possible connections across every component, with an algo
that grew in time linearly with the number of attribute values in the
entire change set. But we only need to calculate them per-component!

This improves cold start loads of prod saas in tools (in the no hammer
case) from 100+ seconds to around 60 for me.