Add resource values panel
This change adds the "ResourceValuesPanel" for looking at a read-only
tree of values corresponding resource of a given component. There are
three states for this panel: non-existence (the component will not have
a resource), empty state (the component does not have a resource) and
populated (the component has a resource). The empty state design
is copied from "ResourcePanel".

At a nuts and bolts level, the new panel is a stripped down version of
the "AttributePanel". In fact, so much has been re-used that some logic
has been moved into a new composable library and likely more logic can
be moved there in the near future.

Signed-off-by: Nick Gerace <nick@systeminit.com>