Do not coerce value for subscriptions
This change ensures that we do not coerce a value when dealing with
subscriptions. This bug was found when trying to subscribe from (to?) a
prop with kind "integer", but the value would be the prop path, which
is a "string". This would result in "NaN" being the coerced value and
then "null" sent to the attributes API. Now, we don't even try to coerce
if we are dealing with a subscription.

Co-authored-by: Paul Stack <paul@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>