feat(web): force election, and retry if leader changed mid req
This pr makes two changes that should increase the reliablity of the
multitab experience:

  1. If lock acquisition has not been detected after 5 seconds, attempt
     to force leader election.
  2. If the leader changes in the middle of a request, that request will
     likely hang. So detect this change using Promise.race and retry the
     call in the case that leader election fails.

Also upgrades sqlite-wasm to 3.50.3-build1, which was required to force
leader election by closing the database and pausing the VFS.

Test this: open lots of tabs. Reboot the leader tab mid MUSPELHEIM.
Muspelheim should still finish in all tabs. Testing the leader election
force is harder since it should only happen when the tab that is
acquiring the lock is in a broken state.