Add new validation failure UI
This change adds a new UI element for failed validations. When a
validation is failing that is not a "requirement" one, we now expand
the area around the input box and indicate within the panel why the
field is failing. This to make the usability flow for both keyboard
and mouse users better. The tooltip has been removed and full message
is now displayed.

This change also ensures that the asterisk is present when the possible
connections box is present.

Signed-off-by: Nick Gerace <nick@systeminit.com>