Refresh actions can run in change sets
When importing resources, it is quite painful to have to apply a change set in order to propagate resource_values to use in downstream components. Similarly, if you know a resource has changed, it is unnecessary to have to force a new change set to get the updated values.  To improve this, the following changes were made: 

In the management function job, if we're running an Import function, and we're NOT on head, we will now dispatch that action job from the change set so that we can set the resource and propagate values as needed.

A new route was added to SDF that handle refresh actions only with the following logic: 
1. If refresh was triggered on head, enqueue the action
2. If refresh was triggered in an open change set (NOT head), check if the component exists on head. If it does, we enqueue the refresh action on head (and upon completion the updates will be replayed back to the change set). If it does NOT exist on head, we create an action, mark it as dispatched, and dispatch it directly from the change set! 

We get two key benefits from this: 
1. A user can import a full suite of resources, and create connections as they go, without having to do the apply/open new change set dance.
2. A user can trigger a refresh from their change set, and see the results flow back, and they can trigger a refresh directly on head without having to open a change set