Overhaul mgmt function UX and special case funcs
This change is a general overhaul to the UX and UI for management funcs
with special case behavior for "import" and "run template" funcs. This
design went through a bunch of iterations, but this one is a good start
towards handling the balance of knowing the state of a management func
(and not caring about whether the function or operations failed) and
being able to debug it with some granularity.

Essentially, this change ensures that the management func tab and
general status tracking for management funcs includes both the func run
and the operator. Originally, this change used the status "bucket" from
the StatusPanel, but given the requirement to view live WsEvents as well
as the need for both FuncRunDetails and ManagementFuncCard to know the
status of the operator upon initial load, using the job state database
table resulted in a cleaner approach.

The status for management funcs is now centralized and is available for
both the management func card and the component details page. It uses
"funcRunStatus" because that function populates the func run status
icons on the Explore page and the individual func run pages. The
management func card now disallows "run template" funcs are they are
special-cased in addition to "import" funcs.

From these changes, we also now allow error hinting via banners. There
is one user of this subsystem now: the management funcs. The hint and
raw error message are provided by checking the func run state and the
job state as well as using the new "message" column for the job state
table. It indicates the reason behind the current state. This is only
used for failing states at the moment and is type guarded.

Additional Changes:
- Add ErrorBanner for use in both ComponentDetails (grid compliant with
  a new "fit-content" usage... thank you @jobelenus) and on the func
  run page
- Add a new "si_id" library within "newhotness" to account for IDs
  without a home that are present in "lib/si-id"
- Add destructive color for newhotness designs within the Tailwind
  colors to prevent hex color sprawl (this affects the existing usages
  of the color in validations land)

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>