feat: updated template generation
This PR implements updated template generation. It introduces the
'si-generate-template' crate, which uses the askama templating library
to generate typescript function code. It uses the 'attributes' syntax 
for setting all properties, uses variable declarations in the management
function for easier maintenance, and sets sources correctly.

If you use AWS components in the template, it will automatically wire in
the credential and the region from the template. 

It uses a 'Name Prefix' field to automatically prefix the name of the
generated components. If there are components with that prefix already
attached via a management edge, the function will fail (preventing
multiple runs with the same prefix.) No attempt is made to make the
template idempotent.