feat(edda): introduce deployment & workspace tasks
This change extends the kinds of tasks that Edda can execute. Prior to
this change, all Edda instances would subscribe to the `EDDA_TASKS`
stream for change set tasks which would be spun up while there was
incoming work for the change set.

With this change, the NATS subject structure is updated in the
`EDDA_TASKS` and `EDDA_REQUESTS` streams to add a "scope" segement,
meaning that a former change sets task message on:

```
edda.tasks.$workspace_id.$change_set_id.$task_kind
```

will now be sent on:

```
edda.tasks.change_set.$workspace_id.$change_set_id.$task_kind
```

The third subject segment is a "scope" segement, which current would be
on of:

- `change_set`: a task to run for a specific change set in a workspace
- `workspace`: a task to run for a specific workspace (and/or across all
  of a workspace's change set depending how would would like to think
  about it)
- `deployment`: a task to run across the entire deployed System
  Initiative system (and/or across all change sets and workspaces on the
  sytem if that's an appropriate approach)

Currently the "deployment" scope is added and listening for (currently)
unstructured requests that have an empty processing implementation. All
the other code works as before. The "deployment" scope is where future
work will add cached module MV generation, which are objects that can be
shared for every workspace and every change set.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>