Bring back ability to add/remove from a view
This change brings back the ability to add components to a view and
to remove components from a view.

Description:

There are a lot of changes to the webworker and surrounding areas, but
the primary change involves cache busting "ViewComponentList". This is
because "ViewComponentList" and other listables are not supposed to be
cache-busted. However, this is the first time where we need to do it
until we come up with something better because, without it, making
changes to components in views will not be reflected in the UI. I tested
with this ~300 components locally, and while the initial load wasn't
great, subsequent use of the components worked well.

Having fixed the reactivity issue, the issue then became ensuring that
the dropdown data was sufficient. Not only did we need to make sure
that components could be added and removed from views (rather than just
tanking 403 errors), but we also needed to make sure that components
would not be orphaned. This results in two new queries that are also
special cased in the cache busting world: "componentsInViews" and
"componentsInOnlyOneView". These are included in a composable that is
used by the context menu to determine what to show. Fortunately,
performance here is solid with around ~300 components.

History:

Originally, this used the "current" view for everything, which made
the initial implementation functional, but very limited and unusable
for adding components to anything other than the default view. That's
because you could have only selected what you saw, so if it was not
in the view yet, you couldn't select it to add to that view. If you were
in the "All Views" mode, you could add it to the default view, but that
was the only way that adding "worked". Removing from a view worked well
because you could remove what you could see, but that was still a
limited experience compared to what this change now includes.

Both of these worked with keyboard shortcuts because we knew the view
that you wanted to add/remove from. Those have been removed as a result
of needing a breakout menu.

Additional Changes:

This change includes shortcut modal changes, including the addition of
the "right click" option. There would not have been any shortcut modal
changes, but since this change originally included shortcuts for the
add/remove functionality, they were already in place.

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Co-authored-by: John Keiser <jkeiser@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>