feat(frigg): introduce deployment objects & index
This change extends the KV store's key structure to include the concept
of "scopes", similar to the updates to Edda's NATS streams. Prior to
this change, the only objects that Figg stored were front end objects to
be shared within a specific workspace and the only indexes were for
change sets within a workspace.

With these changes, the implicit object and index scopes are made
explicit, leaving room for more objects and indexes. All pre-existing
index keys are now change set index keys and are of the form:

```
index.change_set.$workspace_id.$change_set_id
```

Similarily, all pre-existing object keys are now workspace object keys,
as these frontend objects may be shared across multiple change sets--but
always scoped within a workspace. So, these workspace object keys are
now of the form:

```
object.workspace.$workspace_id.$object_kind.$object_id.$object_checksum
```

All Frigg API methods are updated to reflect these more precise
concepts which leaves room to introduce 2 new concepts.

A deployment index key is a singular item which is used for the entire
deployment (and/or for all change sets and all workspaces if that
meaning makes more sense) and is of the form:

```
index.deployment
```

A deployment object key refers to a front end object which can be used
for every change set and every workspace in the deployed System
Initiative system, and is of the form:

```
object.deployment.$object_kind.$object_id.$object_checksum
```

All the pre-exisiting API methods work as before and future work will
use the new deployment method calls to populate this content.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>