feat(dal): default subscriptions for scalar values
Adds a new category node, `DefaultSubscriptionSources`, and a new edge
kind: `DefaultSubscriptionSource` which points from an attribute value
to the new category.

This allows us to calculate default subscriptions based on the
`suggestSources` and `suggestAsSourceFor` maps in `ui_optionals` for a
schema. If a suggested connection matches one of the default
subscription sources unambiguously, we will be able to create that
subscription when the component is created.

This PR implements the initial logic for discovering the default
subscriptions and adds the new category node. Default subscriptions will
not yet be created on component create (the calculation of suggested
sources is expensive and needs to be done ahead of time and cached).

Only scalars supported in this PR. A follow up is coming for arrays,
objects, and maps.