feat: add import and discovery to mcp server, update discovery in clover
This PR adds an import and discovery tool to the MCP server. It also
adds custom import and discovery prompts, which can be accessed as '/'
commands in claude code, and will do more complex import and discovery
processes.

The change to clover updates the logic that transferred incoming
connections to the discovery template to discovered components to
transferring the 'sources' - this means that region and credential are
correctly connected. 

This means an update to every clover component. Yay churn!