Indicate whether Components (and attribute values) have socket connections so we can warn users migration wasn't cleanly possible
This change implements a method `has_socket_connections()` on `WorkspaceSnapshotSelector` which determines whether a component has at least one input socket with a connection, using lower level abstractions to reduce the performance hit of this work.

We include this data in the `ComponentInListMV` and, when true, for each `AttributeValue` we'll attempt to find if the attribute value is bein set by a socket. 

I've also removed `canBeSetBySocket` from the MV as we do not use this data in newhotness.