Resurrect workspace approvals
This change brings back workspace-level approvals and the approvals
subsystem with an eye for the future. The inset modal has been converted
into a box within the flow modal. The notifications system, including
approving from another change set, has been brought back too. Grab a
chair. This ended up being a big one.

Strategy:

The conversion and naming of "InsetApprovalModal" were both intentional.
It is no longer inset, but retains the name temporarily for reviewers
and to cleanly convey intent about what it does. There will be a quick
follow-up change to fix any inconsistent naming, potential
anti-patterns, etc. This name will have to change.

The style of the buttons and inner content of that modal are retained
as well. You may notice that the apply button is the old one. This is
entirely intentional. In the last attempt (PR #6816), part of what
made it a total pain to debug was that the outer "apply" button in
"ApprovalFlowModal.vue" was also taking on the loading, disabling,
labeling, clicking, and styling functionality of the inner
"InsertApprovalModal". We have specifically aimed to not touch the
style since only two workspaces are using this future.

The strategy was essentially to cleanly port that inset modal into
a box within the flow modal, but the key was to ensure that clicking
"withdraw request" and "request approval" would flip between the two
modes. The modal dynamically sizes with a transition when you flip
between modes... and I think it looks cool!

In the long term, we should consider integrating the two Vue components
further (particularly the apply/request button), but for now, this
should be functional, easy to debug, and a clean separation of the
approvals feature from the main flow.

Only One Approval Kind:

This change does not explicitly block other kinds of approvals from
existing in the backend and on the graph, but since the UI does not
expose the ability to customize approvals, it will show all existing
requirements as workspace-level changes. Since there are only two
workspaces using approvals and they are workspace-level approvals, we
are covered.

Feature Flagging:

Due to the nest of pinia stores in the old UI, the feature was flagged
pretty aggressively. In this change, the feature is only flagged in
"ApprovalFlowModal.vue". Why? Well, everything else is driven by a prop
being passed down through components, like "Notifications.vue". That
prop has all change sets with a status of "NeedsApproval". You cannot
request approval for a change set outside of the flow modal, so all
affected components will not be running queries or even loaded into the
UI without the feature enabled.

Testing:

This was tested using a workspace with an owner and a collaborator in
two different browsers. Both components with and without actions were
tested. The flag was enabled/disabled via code by changing the
computed value in "ApprovalFlowModal.vue".

Other Things We Benefit From:

- Added ability to work on another change set from another change set
  using the "newCtx" pattern (will require a helper eventually)
- Added ability to provide a "WORKSPACE" context to anyone who needs
  it in the new UI (this was all @jobelenus who put the now-squashed
  commit in my branch)
- Added ability to navigate to an existing change set with a helper
- Added ability to list and track users in the workspace

Near Future Work:

- Rename and refactor anything related to "inset" language
- Fix the change set rename modal anti-pattern and replace it with
  the "newCtx" pattern (seen in PR #6816)
- Move all types from stores to either "@/api/sdf/dal" or any of the
  new directories (e.g. ActionProposedView, approversForChangeSet)

Context:

This change is a successor to PR #6816 and both were inspired
by PR #6633. This required multiple attempts due to the nature of
converting an inset approval into a regular component in the new UI,
dealing with entirely non-MV content and dealing with workspace-level
objects (e.g. users and the "approvals enabled" flag). This attempt
was successful due to re-learning about the domain and finding all the
dragons, but also taking a commit-by-commit, Vue component-by-component
approach to a compiling "newhotness" UI with additional functionality
on an iterative basis. Big commit squash party!

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>